Create ADPL ASSN2
PROBLEM:- A farming community called as Green nest global wants to start an online store because they want to grow their business globally. They are selling onion seeds and other things, and they want to reach more and more customers. They want to make a website and an app to sell their products. The data indicates that they have a low customer retention rate of 20% today. The farming community wants to increase the number of customers to 90%. The purpose of the project is to increase the product sales of the farming community and to reach more farmers, so they can also join the community.

SOLUTION:-We are developing a solution for a common online marketplace website that addresses the issues faced by direct farmers. The platform will include a verified review option, ensuring quality for users by allowing buyers to leave feedback on sellers. Additionally, we will offer the option for customers to buy seeds directly from the website, ensuring that they are purchasing high-quality seeds. The website will also feature information on government schemes, allowing farmers to take advantage of these benefits